pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_project"
version = "0.0.1"
description = "Project for Machine Learning Operations course (02476) by group 102"
authors = [
  { name = "Christian Amtoft Nickelsen", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc==3.66.1",
    "dvc-gdrive==3.0.1",
    "dvc-gs>=3.0.2",
    "fastapi==0.128.0",
    "google-api-python-client>=2.187.0",
    "gsutil>=5.35",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.0",
    "torch>=2.9.1",
    "lightning>=2.3.0",
    "torchmetrics>=1.0.0",
    "tqdm>=4.66.0",
    "typer==0.21.1",
    "uvicorn==0.40.0",
    "pytorch-lightning>=2.6.0",
    "hydra-core>=1.3.2",
    "httpx>=0.28.1",
    "locust>=2.43.1",
    "wandb>=0.23",
    "loguru>=0.7.0",
]

[dependency-groups]
dev = [
    "coverage==7.13.1",
    "invoke==2.2.0",
    "mkdocs-material==9.7.1",
    "mkdocs==1.6.1",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "ruff==0.14.11",
    "mkdocstrings-python==2.0.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cpu", marker = "sys_platform != 'linux' and sys_platform != 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
  { index = "pytorch-cpu", marker = "sys_platform != 'linux' and sys_platform != 'win32'" },
]

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*", "**/tmp*", "*/_remote_module*"]

[tool.coverage.report]
ignore_errors = true
